+++ order/src/main/java/com/high/order/application/dto/internal/kafka/response/OrderSuccessResponse.java
@@ -1,23 +1,17 @@
 package com.high.order.application.dto.internal.kafka.response;
 
 import com.high.order.domain.entity.Order;
-import java.util.List;
 import java.util.UUID;
 
 public record OrderSuccessResponse(
     UUID orderId,
-    UUID sagaId,
-    UUID couponId,
-    UUID ordererId,
-    List<OrderItemSuccessResponse> itemList
+    UUID sagaId
+
 ) {
-    public static OrderSuccessResponse of(Order order, UUID sagaId, UUID ordererId) {
+    public static OrderSuccessResponse of(Order order, UUID sagaId) {
         return new OrderSuccessResponse(
             order.getOrderId(),
-            sagaId,
-            order.getCouponId(),
-            ordererId,
-            order.getOrderItems().stream().map(OrderItemSuccessResponse::from).toList()
+            sagaId
         );
     }
 }
