+++ order/src/main/java/com/high/order/application/service/OrderServiceV2.java
@@ -92,7 +92,7 @@ public class OrderServiceV2 {
         );
         Order savedOrder = orderRepository.save(order);
 
-        return OrderSuccessResponse.of(order, request.sagaId(), request.ordererId());
+        return OrderSuccessResponse.of(order, request.sagaId());
     }
